requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
discord==1.7.3
discord.py==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
threadpoolctl==2.1.0
urllib3==1.26.3
yarl==1.6.3
python-dotenv==0.19.1